+++ R/summary.R
@@ -41,25 +41,30 @@ setMethod("summary", signature(object="vegtable"),
             cat("## Metadata", "\n")
             if(length(object@description) > 0) {
                 for(i in names(object@description)) {
-                    cat(i, ": ", object@description[i], sep="", "\n")
+                    cat("   ", i, ": ", object@description[i], sep="", "\n")
                 }
             }
-            cat("object size:", format(object.size(object), units=units),
+            cat("   object size:", format(object.size(object), units=units),
                     sep=" ", "\n")
-            cat("validity:", validObject(object), "\n")
+            cat("   validity:", validObject(object), "\n")
             cat("\n")
             # Content of some slots
             cat("## Content", "\n")
-            cat("number of plots:", nrow(object@header), sep=" ", "\n")
-            cat("variables in header:", ncol(object@header), sep=" ", "\n")
-            cat("number of relations:", length(object@relations), sep=" ", "\n")
+            cat("   number of plots:", nrow(object@header), sep=" ", "\n")
+			cat("   plots with records:",
+					length(unique(object@samples$ReleveID)), sep=" ", "\n")
+			cat("   variables in header:", ncol(object@header), sep=" ", "\n")
+            cat("   number of relations:", length(object@relations), sep=" ",
+					"\n")
             cat("\n")
             # Content of species list
             cat("## Taxonomic List", "\n")
-            cat("taxon names:", nrow(object@species@taxonNames), sep=" ", "\n")
-            cat("taxon concepts:", nrow(object@species@taxonRelations), sep=" ",
+            cat("   taxon names:", nrow(object@species@taxonNames), sep=" ",
+					"\n")
+            cat("   taxon concepts:", nrow(object@species@taxonRelations),
+					sep=" ",
                     "\n")
-            cat("validity:", validObject(object@species), sep=" ", "\n")
+            cat("   validity:", validObject(object@species), sep=" ", "\n")
             cat("\n")
         }
 )