+++ R/vegtable-class.R
@@ -52,20 +52,26 @@ setClass("vegtable",
                 relations=list(),
 				coverconvert=new("coverconvert")),
 		validity=function(object) {
-            # Mandatory names
-            if(any(!c("ReleveID","TaxonUsageID") %in% colnames(object@samples)))
-                return("Columns 'ReleveID' and 'TaxonUsageID' are mandatory in slot 'samples'")
+            # Mandatory names in samples (only for objects with records)
+            if(nrow(object@samples) > 0 & any(!c("ReleveID","TaxonUsageID") %in%
+							colnames(object@samples)))
+                return(paste("Columns 'ReleveID' and 'TaxonUsageID' are",
+								"mandatory in slot 'samples'"))
+			# TODO: Use a better message (showing missmatching names)
 			for(i in names(object@layers)) {
 				if(!i %in% colnames(object@samples))
-					return(paste0("Layers of '", i, "' not included in slot 'samples'"))
+					return(paste0("Layers of '", i,
+									"' not included in slot 'samples'"))
 				if(!i %in% colnames(object@layers[[i]]))
-					return(paste0("Column '", i, "' is mandatory in layer table '", i, "'"))
+					return(paste0("Column '", i,
+									"' is mandatory in layer table '", i, "'"))
 			}
 			if(!"ReleveID" %in% colnames(object@header))
                 return("Column 'ReleveID' is mandatory in slot 'header'")
             for(i in names(object@relations)) {
                 if(!i %in% colnames(object@header))
-                    return(paste0("Relation '", i, "' not included in slot 'header'"))
+                    return(paste0("Relation '", i,
+									"' not included in slot 'header'"))
                 if(!i %in% colnames(object@relations[[i]]))
                     return(paste0("Column '", i, "' is mandatory in relation '", i, "'"))
 				if(any(!object@header[,i] %in% object@relations[[i]][,i] & !is.na(object@header[,i])))
@@ -75,9 +81,7 @@ setClass("vegtable",
             # Mandatory links
             if(!all(object@samples$ReleveID %in% object@header$ReleveID))
                 return("Some releve IDs from slot 'samples' are missing in slot 'header'")
-            if(!all(object@header$ReleveID %in% object@samples$ReleveID))
-                return("Some releve IDs from slot 'header' are missing in slot 'samples'")
-            if(any(is.na(object@samples$TaxonUsageID)))
+			if(any(is.na(object@samples$TaxonUsageID)))
 				return("NAs are not allowed in 'TaxonUsageID' at slot 'samples'.")
 			if(any(!object@samples$TaxonUsageID %in%
 							object@species@taxonNames$TaxonUsageID))